+++ backend/canisters/user/impl/src/queries/search_messages.rs
@@ -2,7 +2,6 @@ use crate::guards::caller_is_owner;
 use crate::{RuntimeState, read_state};
 use canister_api_macros::query;
 use oc_error_codes::OCErrorCode;
-use search::simple::Query;
 use std::collections::HashSet;
 use types::{MessageIndex, OCResult};
 use user_canister::search_messages::{Response::*, *};
@@ -30,10 +29,10 @@ fn search_messages_impl(args: Args, state: &RuntimeState) -> OCResult<SuccessRes
     }
 
     let direct_chat = state.data.direct_chats.get_or_err(&args.user_id.into())?;
-    let query = Query::new(&args.search_term);
-    let matches = direct_chat
-        .events
-        .search_messages(MessageIndex::default(), query, HashSet::new(), args.max_results);
+    let matches =
+        direct_chat
+            .events
+            .search_messages(MessageIndex::default(), &args.search_term, &HashSet::new(), args.max_results);
 
     Ok(SuccessResult { matches })
 }